Add support for gatsby's `pathPrefix` option
If `pathPrefix` is defined in Gatsby's config, it will be automatically
prepended to all local `src`, `href`, etc paths. This causes the SRI
plugin's search+replace to fail if it uses only the local filesystem
path to derive a search pattern.

This patch allows the plugin to take its own `pathPrefix` option which
will then be prefixed to all search/replace patterns.

c.f.
https://www.gatsbyjs.com/docs/how-to/previews-deploys-hosting/path-prefix/